refactor: atomsの階層を更に分割
