+++ src/components/atoms/image/mainImage.js
