+++ src/components/atoms/image/mainImage.module.scss
