+++ src/components/atoms/thubnail/largeThumbnail.js
