+++ src/components/atoms/thubnail/largeThumbnail.module.scss
