+++ src/components/atoms/thubnail/smallThumbnail.js
