+++ src/components/atoms/thubnail/smallThumbnail.module.scss
