+++ src/components/ecosystem/member/inui_window.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import IconButton from '../../atoms/icon_button';
+import IconButton from '../../atoms/button/icon_button';
 import CharMessage from '../../molecules/charMessage';
 import styles from '../window_layout.module.scss';
