+++ src/components/molecules/charMessage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './charMessage.module.scss';
 
-import CharImage from '../atoms/charImage';
+import CharImage from '../atoms/image/charImage';
 import Balloon from '../atoms/balloon';
 
 /**
